requirements.txt
# Core dependencies
numpy==1.26.4
Pillow==10.3.0

# HuggingFace
datasets==2.19.0
huggingface-hub==0.23.0

# AWS S3
boto3==1.34.110

# Optional
requests==2.31.0
